+++ packages/patrol_cli/test/features/drive/device_finder_test.dart
@@ -3,7 +3,7 @@ import 'package:patrol_cli/src/features/devices/device_finder.dart';
 import 'package:patrol_cli/src/features/drive/device.dart';
 import 'package:test/test.dart';
 
-import '../../command_runner_test.dart';
+import '../../mocks.dart';
 import 'fixures/devices.dart';
 
 void main() {
